# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: SoloShot AI API
  version: 1.0.0
  description: >-
    Contract-first API for SoloShot AI. All normalized coordinates use the [0,1]
    range with a top-left origin; x grows right and y grows down. W3 adds a
    short-lived, capability-secured H5 to iOS handoff surface.
servers:
  - url: http://localhost:8000
    description: Local development
...
paths:
...
  /api/v1/sessions:
    post:
      operationId: createSession
      tags: [sessions]
      summary: Create a versioned SoloShot session
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSessionRequest"
      responses:
        "201":
          $ref: "#/components/responses/SessionCreated"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/sessions/{session_id}:
    parameters:
      - $ref: "#/components/parameters/SessionId"
    get:
      operationId: getSession
      tags: [sessions]
      summary: Retrieve a session
      responses:
        "200":
          $ref: "#/components/responses/SessionOk"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteSession
      tags: [sessions]
      summary: Delete a session and schedule owned temporary media for deletion
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          $ref: "#/components/responses/Deleted"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/sessions/{session_id}/capture-consent:
    post:
      operationId: recordCaptureConsent
      tags: [sessions]
      summary: Record iOS consent for temporary capture upload and optional external AI evaluation
      parameters:
        - $ref: "#/components/parameters/SessionId"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/HandoffClaimToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CaptureConsentRequest"
      responses:
        "200":
          $ref: "#/components/responses/CaptureConsentOk"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/v1/media/uploads:
    post:
      operationId: createMediaUpload
      tags: [media]
      summary: Create a short-lived signed upload for session-owned image media
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/OptionalHandoffClaimToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMediaUploadRequest"
      responses:
        "201":
          $ref: "#/components/responses/MediaUploadCreated"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/v1/media/uploads/{media_asset_id}/complete:
    post:
      operationId: completeMediaUpload
      tags: [media]
      summary: Verify an uploaded object and mark the media ready
      parameters:
        - $ref: "#/components/parameters/MediaAssetId"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/OptionalHandoffClaimToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompleteMediaUploadRequest"
      responses:
        "200":
          $ref: "#/components/responses/MediaAssetOk"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/v1/media/{media_asset_id}/access:
    get:
      operationId: getMediaAccess
      tags: [media]
      summary: Create a short-lived read URL for ready session-owned media
      parameters:
        - $ref: "#/components/parameters/MediaAssetId"
        - $ref: "#/components/parameters/SessionIdQuery"
      responses:
        "200":
          $ref: "#/components/responses/MediaAccessOk"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/references/analyze:
    post:
      operationId: analyzeReference
      tags: [references]
      summary: Analyze a selected reference subject
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyzeReferenceRequest"
      responses:
        "202":
          $ref: "#/components/responses/ReferenceAnalysisAccepted"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /api/v1/references/adapt:
    post:
      operationId: adaptReference
      tags: [references]
      summary: Adapt reference composition to a current scene
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdaptReferenceRequest"
      responses:
        "202":
          $ref: "#/components/responses/ReferenceAnalysisAccepted"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/v1/references/validate-box:
    post:
      operationId: validateReferenceBox
      tags: [references]
      summary: Validate a normalized subject selection box
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ValidateReferenceBoxRequest"
      responses:
        "200":
          $ref: "#/components/responses/ReferenceBoxValidationOk"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/v1/references/{reference_id}:
    get:
      operationId: getReference
      tags: [references]
      summary: Retrieve reference metadata
      parameters:
        - $ref: "#/components/parameters/ReferenceId"
      responses:
        "200":
          $ref: "#/components/responses/ReferenceOk"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/agent/runs:
    post:
      operationId: createAgentRun
      tags: [agent]
      summary: Start a versioned Agent run
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAgentRunRequest"
      responses:
        "202":
          $ref: "#/components/responses/AgentRunAccepted"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /api/v1/agent/runs/{run_id}:
    get:
      operationId: getAgentRun
      tags: [agent]
      summary: Retrieve Agent run state
      parameters:
        - $ref: "#/components/parameters/RunId"
      responses:
        "200":
          $ref: "#/components/responses/AgentRunOk"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/agent/runs/{run_id}/continue:
    post:
      operationId: continueAgentRun
      tags: [agent]
      summary: Continue a multi-round Agent run
      parameters:
        - $ref: "#/components/parameters/RunId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContinueAgentRunRequest"
      responses:
        "202":
          $ref: "#/components/responses/AgentRunAccepted"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /api/v1/agent/runs/{run_id}/trace:
    get:
      operationId: getAgentRunTrace
      tags: [agent]
      summary: Retrieve the safe Agent and Skill trace
      parameters:
        - $ref: "#/components/parameters/RunId"
      responses:
        "200":
          $ref: "#/components/responses/AgentTraceOk"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/skills/{skill_name}/invoke:
    post:
      operationId: invokeSkill
      tags: [skills]
      summary: Invoke one registered Skill through the backend
      parameters:
        - $ref: "#/components/parameters/SkillName"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvokeSkillRequest"
      responses:
        "202":
          $ref: "#/components/responses/SkillRunAccepted"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
        "409":
          $ref: "#/components/responses/Conflict"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /api/v1/captures:
    post:
      operationId: createCapture
      tags: [captures]
      summary: Register an uploaded capture
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/OptionalHandoffClaimToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCaptureRequest"
      responses:
        "201":
          $ref: "#/components/responses/CaptureCreated"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/v1/captures/{capture_id}:
    get:
      operationId: getCapture
      tags: [captures]
      summary: Retrieve capture processing state
      parameters:
        - $ref: "#/components/parameters/CaptureId"
      responses:
        "200":
          $ref: "#/components/responses/CaptureOk"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/captures/{capture_id}/select-frame:
    post:
      operationId: selectCaptureFrame
      tags: [captures]
      summary: Select one candidate frame
      parameters:
        - $ref: "#/components/parameters/CaptureId"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/OptionalHandoffClaimToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SelectFrameRequest"
      responses:
        "200":
          $ref: "#/components/responses/CaptureOk"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/evaluations:
    post:
      operationId: createEvaluation
      tags: [evaluations]
      summary: Compare a capture with its reference and plan
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/OptionalHandoffClaimToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEvaluationRequest"
      responses:
        "202":
          $ref: "#/components/responses/EvaluationAccepted"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /api/v1/handoffs:
    post:
      operationId: createHandoff
      tags: [handoffs]
      summary: Create a short-lived handoff code
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateHandoffRequest"
      responses:
        "200":
          $ref: "#/components/responses/HandoffCreated"
        "201":
          $ref: "#/components/responses/HandoffCreated"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/v1/handoffs/{code}:
    parameters:
      - $ref: "#/components/parameters/HandoffCode"
    get:
      operationId: getHandoff
      tags: [handoffs]
      summary: Retrieve safe handoff metadata
      responses:
        "200":
          $ref: "#/components/responses/HandoffOk"
        "404":
          $ref: "#/components/responses/NotFound"
        "410":
          $ref: "#/components/responses/Gone"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
    delete:
      operationId: revokeHandoff
      tags: [handoffs]
      summary: Revoke a handoff
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/HandoffManagementToken"
      responses:
        "200":
          $ref: "#/components/responses/HandoffOk"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "410":
          $ref: "#/components/responses/Gone"
  /api/v1/handoffs/{code}/claim:
    post:
      operationId: claimHandoff
      tags: [handoffs]
      summary: Claim a handoff once
      parameters:
        - $ref: "#/components/parameters/HandoffCode"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClaimHandoffRequest"
      responses:
        "200":
          $ref: "#/components/responses/HandoffClaimed"
        "409":
          $ref: "#/components/responses/Conflict"
        "410":
          $ref: "#/components/responses/Gone"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /api/v1/handoffs/{code}/complete:
    post:
      operationId: completeHandoff
      tags: [handoffs]
      summary: Mark a claimed handoff complete
      parameters:
        - $ref: "#/components/parameters/HandoffCode"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/HandoffClaimToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompleteHandoffRequest"
      responses:
        "200":
          $ref: "#/components/responses/HandoffOk"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
        "410":
          $ref: "#/components/responses/Gone"
  /api/v1/posts/render:
    post:
      operationId: renderPost
      tags: [content]
      summary: Queue deterministic Before/After and 9:16 content rendering
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RenderPostRequest"
      responses:
        "202":
          $ref: "#/components/responses/PostAccepted"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /api/v1/posts/{post_id}:
    get:
      operationId: getPost
      tags: [content]
      summary: Retrieve content render state
      parameters:
        - $ref: "#/components/parameters/PostId"
      responses:
        "200":
          $ref: "#/components/responses/PostOk"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/events/batch:
    post:
      operationId: recordEventBatch
      tags: [analytics]
      summary: Record a deduplicated batch of privacy-safe events
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventBatchRequest"
      responses:
        "202":
          $ref: "#/components/responses/EventBatchAccepted"
        "422":
          $ref: "#/components/responses/ValidationError"
...
components:
  securitySchemes:
    internalBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque
...
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Opaque key used to deduplicate retryable writes
      schema:
        type: string
        minLength: 8
        maxLength: 128
    SessionId:
      name: session_id
      in: path
      required: true
      schema:
        type: string
        pattern: ^ss_[A-Za-z0-9_-]+$
...
    MediaAssetId:
      name: media_asset_id
      in: path
      required: true
      schema:
        type: string
        pattern: ^media_[A-Za-z0-9_-]+$
...
    HandoffClaimToken:
      name: X-Handoff-Claim-Token
      in: header
      required: true
      description: Opaque claimant capability bound to one client instance
      schema:
        type: string
        minLength: 32
        maxLength: 2048
    OptionalHandoffClaimToken:
      name: X-Handoff-Claim-Token
      in: header
      required: false
      description: Required for capture writes after a Handoff has been completed by iOS
      schema:
        type: string
        minLength: 32
        maxLength: 2048
...
  schemas:
...
    CreateSessionRequest:
      type: object
      required: [schema_version, source_channel, mode, user_constraints]
      properties:
        schema_version:
          type: string
          const: "1.0"
        source_channel:
          $ref: ./schemas/common.schema.json#/$defs/SourceChannel
        mode:
          type: string
          enum: [original_replication, scene_adaptation]
        user_constraints:
          $ref: ./schemas/common.schema.json#/$defs/UserConstraints
        external_ai_consent:
          type: boolean
          default: false
      additionalProperties: false
    CreateMediaUploadRequest:
      type: object
      required: [schema_version, session_id, purpose, content_type, byte_size, sha256]
      properties:
        schema_version:
          type: string
          const: "1.0"
        session_id:
          type: string
          pattern: ^ss_[A-Za-z0-9_-]+$
        purpose:
          $ref: ./schemas/media.schema.json#/$defs/MediaPurpose
        content_type:
          type: string
          enum: [image/jpeg, image/png, image/webp]
        byte_size:
          type: integer
          minimum: 1
          maximum: 8000000
        sha256:
          type: string
          pattern: ^[a-f0-9]{64}$
      additionalProperties: false
    CaptureConsentRequest:
      type: object
      required: [schema_version, capture_upload_consent, external_ai_consent]
      properties:
        schema_version:
          type: string
          const: "1.0"
        capture_upload_consent:
          type: boolean
          description: Must be true; the service rejects false because capture upload requires explicit consent.
        external_ai_consent:
          type: boolean
      additionalProperties: false
...
  responses:
    SessionCreated:
      description: Session created
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/SessionResponse"
    SessionOk:
      description: Session retrieved
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/SessionResponse"
    CaptureConsentOk:
      description: Capture consent recorded
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/CaptureConsentResponse"
...
    Deleted:
      description: Resource deletion or revocation accepted
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/DeletedResponse"
...
    BadRequest:
      description: Invalid request
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication required
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Resource state conflicts with this operation
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    ValidationError:
      description: Input failed schema or domain validation
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: Unexpected server error
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...